Parameter API refactoring in tests & examples
