+++ examples/managed-client-simple-webapp/src/main/java/org/glassfish/jersey/examples/managedclientsimple/resources/ClientResource.java
@@ -7,6 +7,7 @@
  *
  * SPDX-License-Identifier: BSD-3-Clause
  */
+// Portions Copyright [2018] [Payara Foundation and/or its affiliates]
 
 package org.glassfish.jersey.examples.managedclientsimple.resources;
 
@@ -19,7 +20,7 @@ import javax.ws.rs.client.WebTarget;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
-import org.glassfish.jersey.server.Uri;
+import org.glassfish.jersey.Uri;
 
 /**
  * A resource which use managed client injected by {@link org.glassfish.jersey.server.Uri @Uri annotation} to query
