+++ examples/open-tracing/src/main/java/org/glassfish/jersey/examples/opentracing/TracedResource.java
@@ -7,6 +7,7 @@
  *
  * SPDX-License-Identifier: BSD-3-Clause
  */
+// Portions Copyright [2018] [Payara Foundation and/or its affiliates]
 
 package org.glassfish.jersey.examples.opentracing;
 
@@ -26,7 +27,7 @@ import javax.ws.rs.core.Response;
 import org.glassfish.jersey.opentracing.OpenTracingFeature;
 import org.glassfish.jersey.opentracing.OpenTracingUtils;
 import org.glassfish.jersey.server.ManagedAsync;
-import org.glassfish.jersey.server.Uri;
+import org.glassfish.jersey.Uri;
 
 import io.opentracing.Span;
