+++ examples/rx-client-webapp/src/main/java/org/glassfish/jersey/examples/rx/agent/SyncAgentResource.java
@@ -7,6 +7,7 @@
  *
  * SPDX-License-Identifier: BSD-3-Clause
  */
+// Portions Copyright [2018] [Payara Foundation and/or its affiliates]
 
 package org.glassfish.jersey.examples.rx.agent;
 
@@ -29,7 +30,7 @@ import org.glassfish.jersey.examples.rx.domain.Calculation;
 import org.glassfish.jersey.examples.rx.domain.Destination;
 import org.glassfish.jersey.examples.rx.domain.Forecast;
 import org.glassfish.jersey.examples.rx.domain.Recommendation;
-import org.glassfish.jersey.server.Uri;
+import org.glassfish.jersey.Uri;
 
 /**
  * Obtain information about visited (destination) and recommended (destination, forecast, price) places for "Sync" user. Uses
